Add gosu and entrypoint to jessie-curl
The idea is to provide the image with the gosu command.

By default, additional user is created with name 'phantom'
with the UID=1000. This user is used in entrypoint script to execute
provided commands.

The following commands are added to the image:
  - gosu

The following files are added to the image:
  - docker-entrypoint.sh